+++ include/fi_enosys.h
@@ -33,7 +33,7 @@
 #ifndef _FI_ENOSYS_H_
 #define _FI_ENOSYS_H_
 
-#include <fi_osd.h>
+#include <ofi_osd.h>
 
 #include <rdma/fabric.h>
 #include <rdma/fi_atomic.h>
