+++ include/windows/osd.h
@@ -33,7 +33,7 @@
 #include <rdma/fi_errno.h>
 #include <rdma/fabric.h>
 
-#include <fi_osd.h>
+#include <ofi_osd.h>
 
 #ifdef __cplusplus
 extern "C" {
