+++ prov/netdir/src/netdir_buf.h
@@ -38,7 +38,7 @@
 #include "rdma/fabric.h"
 
 #include "fi.h"
-#include "fi_osd.h"
+#include "ofi_osd.h"
 
 #include "netdir.h"
