+++ prov/sockets/src/sock_fabric.c
@@ -48,7 +48,7 @@
 #endif
 
 #include "prov.h"
-#include "fi_osd.h"
+#include "ofi_osd.h"
 
 #include "sock.h"
 #include "sock_util.h"
