+++ src/mem.c
@@ -37,7 +37,7 @@
 #include <ctype.h>
 #include <inttypes.h>
 
-#include <fi_osd.h>
+#include <ofi_osd.h>
 #include <fi.h>
 #include <rdma/fi_errno.h>
 #include <ofi_mem.h>
