+++ util/info.c
@@ -34,7 +34,7 @@
 #include <string.h>
 #include <getopt.h>
 
-#include <fi_osd.h>
+#include <ofi_osd.h>
 
 #include <rdma/fabric.h>
 #include <rdma/fi_errno.h>
